Half precision, empty_cache

+++ big_sleep/clip.py
@@ -338,7 +338,7 @@ class LayerNorm(nn.LayerNorm):
 
     def forward(self, x: torch.Tensor):
         orig_type = x.dtype
-        ret = super().forward(x.type(torch.float32))
+        ret = super().forward(x.type(self.weight.dtype))
         return ret.type(orig_type)
